+++ src/App.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 
 import Auth from './pages/Auth';
-import Error from './pages/Error';
 import useRedirect from './hooks/useRedirect';
 
 function App() {
@@ -11,7 +10,6 @@ function App() {
     <div className="App">
       <Routes>
         <Route index element={<Auth />} />
-        <Route path="/error" element={<Error />} />
       </Routes>
     </div>
   );
