+++ src/hooks/useRedirect.ts
@@ -1,14 +1,13 @@
-import { useNavigate, useLocation } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import { useEffect } from 'react';
 
 const useRedirect = () => {
-  const { pathname } = useLocation();
   const navigate = useNavigate();
   const isToken = localStorage.getItem('token');
 
   useEffect(() => {
     if (isToken) navigate('/todo');
-    else if (!isToken && pathname === '/todo') navigate('/');
+    else navigate('/');
   }, [isToken, navigate]);
 };
